+++ App.xaml
@@ -2,8 +2,8 @@
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:local="clr-namespace:TODO"
-             xmlns:viewModel="clr-namespace:TODO.MVVM.ViewModel"
-             xmlns:view="clr-namespace:TODO.MVVM.View"
+             xmlns:viewModel="clr-namespace:TODO.ViewModel"
+             xmlns:view="clr-namespace:TODO.View"
              StartupUri="MainWindow.xaml">
     <Application.Resources>
         <ResourceDictionary>
