+++ Converter/NullToBoolConverter.cs
@@ -1,13 +1,16 @@
 ﻿using System.Globalization;
 using System.Windows.Data;
 
-namespace TODO.Utils
+namespace TODO.Converter
 {
-    public class BooleanToTextConverter : IValueConverter
+    public class NullToBoolConverter : IValueConverter
     {
+        public bool NullValue { get; set; } = false;
+        public bool NonNullValue { get; set; } = true;
+
         public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
         {
-            return (bool)value ? "Completed" : "Complete";
+            return value == null ? NullValue : NonNullValue;
         }
 
         public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
