+++ Converter/StringToVisibilityConverter.cs
@@ -1,0 +1,21 @@
+﻿using System.Globalization;
+using System.Windows.Data;
+using System.Windows;
+
+namespace TODO.Converter
+{
+    public class StringToVisibilityConverter : IValueConverter
+    {
+        public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
+        {
+            return string.IsNullOrEmpty(value?.ToString())
+                ? Visibility.Collapsed
+                : Visibility.Visible;
+        }
+
+        public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
+        {
+            throw new NotImplementedException();
+        }
+    }
+}
