+++ MainWindow.xaml
@@ -4,7 +4,7 @@
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         xmlns:local="clr-namespace:TODO"
-        xmlns:viewModel="clr-namespace:TODO.MVVM.ViewModel"
+        xmlns:viewModel="clr-namespace:TODO.ViewModel"
         xmlns:utils="clr-namespace:TODO.Utils"
         mc:Ignorable="d"
         Height="600" Width="920"
