+++ Themes/CardThemes/TodoCard.xaml
@@ -1,12 +1,12 @@
 ﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
                     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-                    xmlns:local="clr-namespace:TODO.Utils">
+                    xmlns:converter="clr-namespace:TODO.Converter">
 
     <ResourceDictionary.MergedDictionaries>
         <ResourceDictionary Source="/Themes/CheckBoxThemes/ModernCheckBox.xaml" />
     </ResourceDictionary.MergedDictionaries>
 
-    <local:PriorityToColorConverter x:Key="PriorityToColorConverter" />
+    <converter:PriorityToColorConverter x:Key="PriorityToColorConverter" />
 
     <DataTemplate x:Key="TodoCardTemplate">
         <Button Command="{Binding DataContext.OpenEditWindowCommand, RelativeSource={RelativeSource AncestorType=UserControl}}"
