--- Utils/NullToBoolConverter.cs
@@ -1,26 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Globalization;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows.Data;
-
-namespace TODO.Utils
-{
-    public class NullToBoolConverter : IValueConverter
-    {
-        public bool NullValue { get; set; } = false;
-        public bool NonNullValue { get; set; } = true;
-
-        public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
-        {
-            return value == null ? NullValue : NonNullValue;
-        }
-
-        public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
-        {
-            throw new NotImplementedException();
-        }
-    }
-}
