+++ View/HomeView.xaml
@@ -1,9 +1,9 @@
-﻿<UserControl x:Class="TODO.MVVM.View.HomeView"
+﻿<UserControl x:Class="TODO.View.HomeView"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:vm="clr-namespace:TODO.MVVM.ViewModel"
+             xmlns:vm="clr-namespace:TODO.ViewModel"
              mc:Ignorable="d" 
              d:DesignHeight="450" d:DesignWidth="800"
              Background="Transparent">
